requirements.txt
# Core ML & YOLOv5
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# YOLOv5 dependencies
ultralytics==8.0.196
opencv-python==4.8.1.78
matplotlib==3.7.2
pandas==1.5.3
seaborn==0.12.2
scipy==1.10.1

# OCR
easyocr==1.7.1
Pillow==10.4.0

# Utils
tqdm==4.66.1
pyyaml==6.0.1
python-box
ensure
joblib
gdown
kagglehub
scikit-learn
-e . 